requirements_dev.txt
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
blessed==1.19.1
cached-property==1.5.2
cachier==1.5.4
cashier==1.3
cement==2.8.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.3
configparser==3.8.1
cryptography==37.0.3
ctap-keyring-device==1.0.6
Deprecated==1.2.13
docopt==0.6.2
elasticsearch==7.13.0
elasticsearch-dsl==7.4.0
fido2==0.9.3
flake8==4.0.1
future==0.16.0
gimme-aws-creds==2.4.4
humanize==4.2.0
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
jeepney==0.8.0
jmespath==0.10.0
jsonschema==3.2.0
keyring==23.6.0
mccabe==0.6.1
okta==0.0.4
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
pathtools==0.1.2
pluggy==1.0.0
portalocker==2.4.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.1
python-dotenv==0.20.0
PyYAML==5.4.1
redis==4.3.3
requests==2.26.0
SecretStorage==3.3.2
semantic-version==2.8.5
sentry-sdk==1.5.12
six==1.14.0
slack-bolt==1.14.0
slack-sdk==3.17.0
soupsieve==2.3.2.post1
termcolor==1.1.0
texttable==1.6.4
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
watchdog==2.1.8
wcwidth==0.1.9
websocket-client==0.59.0
wrapt==1.14.1
zipp==3.8.0
